Moved version.p where it should be
